adjust signature of from_bytes_rounded_up

+++ lib/types/src/units.rs
@@ -55,8 +55,8 @@ impl Pages {
 
     /// Calculate the number of pages needed to hold the given number of bytes,
     /// rounding up to include any partial page.
-    pub fn from_bytes_rounded_up(bytes: u32) -> Self {
-        Self(bytes.div_ceil(WASM_PAGE_SIZE as u32))
+    pub fn from_bytes_rounded_up(bytes: u64) -> Self {
+        Self(bytes.div_ceil(WASM_PAGE_SIZE as u64) as u32)
     }
 }
